requirements.txt
google_api_python_client==2.54.0
google_auth_oauthlib==0.5.1
pandas==1.1.5
protobuf==4.21.3
pygsheets==2.0.5
